+++ packages/lex_gen/lib/src/services/fmt/lex_packages_generator.dart
@@ -47,7 +47,7 @@ List<LexPackage> generateLexPackagesForService(
   return lexPackages;
 }
 
-List<LexPackage> generateLexPackages(final List<LexType> types) {
+List<LexPackage> generateLexPackages(final List<GeneratableType> types) {
   final packages = <String, Set<String>>{};
   for (final type in types) {
     if (type.isShouldNotBeGenerated()) continue;