+++ packages/lex_gen/lib/src/services/object/lex_input.dart
@@ -7,7 +7,7 @@ import 'lex_property.dart';
 import 'lex_type.dart';
 import 'utils.dart';
 
-final class LexInput extends LexType {
+final class LexInput extends GeneratableType {
   @override
   final String lexiconId;
   @override