+++ packages/lex_gen/lib/src/services/object/lex_known_values.dart
@@ -7,7 +7,7 @@ import '../../utils.dart';
 import '../rule.dart' as rule;
 import 'lex_type.dart';
 
-final class LexKnownValues extends LexType {
+final class LexKnownValues extends GeneratableType {
   @override
   final String lexiconId;
   @override