+++ packages/lex_gen/lib/src/services/object/lex_object.dart
@@ -9,7 +9,7 @@ import 'lex_property.dart';
 import 'lex_type.dart';
 import 'utils.dart';
 
-final class LexObject extends LexType {
+final class LexObject extends GeneratableType {
   @override
   final String lexiconId;
   @override