+++ packages/lex_gen/lib/src/services/object/lex_output.dart
@@ -7,7 +7,7 @@ import 'lex_property.dart';
 import 'lex_type.dart';
 import 'utils.dart';
 
-final class LexOutput extends LexType {
+final class LexOutput extends GeneratableType {
   @override
   final String lexiconId;
   @override